+++ combat/forms.py
@@ -15,15 +15,16 @@ class SnippetForm(forms.ModelForm):
     def save(self, commit=True):
         instance = super(SnippetForm, self).save(commit=False)
 
-        path = 'snippet/{0}/{1}/{2}.{3}'.format(
-            instance.contestant.valid_name(),
-            instance.quiz.valid_name(),
-            instance.valid_name(),
-            'scala' if 'scala' in instance.language else 'py'
-        )
-        default_storage.delete(path)
-        instance.script = default_storage.save(path, ContentFile(instance.body))
-        instance.run_count += 1
+        if self.cleaned_data.get('body', '') != self.initial.get('body', ''):
+            path = 'snippet/{0}/{1}/{2}.{3}'.format(
+                instance.contestant.valid_name(),
+                instance.quiz.valid_name(),
+                instance.valid_name(),
+                'scala' if 'scala' in instance.language else 'py'
+            )
+            default_storage.delete(path)
+            instance.script = default_storage.save(path, ContentFile(instance.body))
+            instance.run_count += 1
         instance.is_running = True
         instance.last_run = timezone.now()
         if commit:
